SelectComponent fixed
The SelectComponent wasn't updating the store values passed to it.
Removing the modifier seemed to do the trick. The test has been updated
to reflect the disabled status on the wrapper itself, which is then
passed down to the element-plus components. Calling the methods directly
on the wrapper was the reason why the test was failing. So we're
avoiding simulating clicks on those components by first checking if the
wrapper has a disabled prop.